work on Mac compilation

+++ compile/macRelink.py
@@ -10,29 +10,6 @@ import sys, os, glob, subprocess, shutil
 def Usage():
     print("\n\tUsage: python "+sys.argv[0]+" <binary-dir>\n")
     sys.exit()
-    
-def MakeByte2str(arg):
-    '''Convert output from subprocess pipes (bytes) to str (unicode) in Python 3.
-    In Python 2: Leaves output alone (already str). 
-    Leaves stuff of other types alone (including unicode in Py2)
-    Works recursively for string-like stuff in nested loops and tuples.
-
-    typical use::
-
-        out = MakeByte2str(out)
-
-    or::
-
-        out,err = MakeByte2str(s.communicate())
-    
-    '''
-    if isinstance(arg,str): return arg
-    if isinstance(arg,bytes): return arg.decode()
-    if isinstance(arg,list):
-        return [MakeByte2str(i) for i in arg]
-    if isinstance(arg,tuple):
-        return tuple([MakeByte2str(i) for i in arg])
-    return arg
 
 if __name__ == '__main__':
     if len(sys.argv) == 2:
@@ -49,8 +26,9 @@ if __name__ == '__main__':
     ignorelist = []
     for f in fileList:
         cmd = ['otool','-L',f]
-        s = subprocess.Popen(cmd,stdout=subprocess.PIPE,stderr=subprocess.PIPE)
-        out,err = MakeByte2str(s.communicate())
+        s = subprocess.Popen(cmd,encoding='UTF-8',
+                    stdout=subprocess.PIPE,stderr=subprocess.PIPE)
+        out,err = s.communicate()
         for i in out.split('\n')[1:]: 
             if not i: continue
             lib = i.split()[0]
@@ -71,10 +49,18 @@ if __name__ == '__main__':
     for key in libs:
         newkey = os.path.join('@rpath',os.path.split(key)[1])
         print('Fixing',key,'to',newkey)
-        shutil.copy(key,dirloc)
+        if os.path.exists(key):
+            shutil.copy(key,dirloc)
+        else:
+            print('skipping copy of',key)
         for f in libs[key]:
             print('\t',os.path.split(f)[1])
             cmd = ["install_name_tool","-change",key,newkey,f]
-            #print(cmd)
-            s = subprocess.Popen(cmd,stdout=subprocess.PIPE,stderr=subprocess.PIPE)
-            out,err = MakeByte2str(s.communicate())
+            print(' '.join(cmd))
+            s = subprocess.Popen(cmd,encoding='UTF-8',
+                                     stdout=subprocess.PIPE,
+                                     stderr=subprocess.PIPE)
+            out,err = s.communicate()
+            if err:
+                print(out)
+                print(err)
